requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.1
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
deprecation==2.1.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
fonttools==4.56.0
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
msgpack==1.1.0
multidict==6.2.0
oscrypto==1.3.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
postgrest==0.19.3
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.4.0
pyphen==0.17.2
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-mock==3.14.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.0
realtime==2.4.2
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.4
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
tzlocal==5.3.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
websockets==14.2
xhtml2pdf==0.2.17
yarl==1.18.3
zopfli==0.2.3.post1
